commit for newer dotnet 8 version, and project rename to bring in line with internal project
