--- PairingTest.sln
@@ -1,31 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.2.32616.157
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PairingTest", "PairingTest\PairingTest.csproj", "{CBE29D97-3942-407F-AF9D-8DA241BD0DEA}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "PairingTest.Tests", "PairingTest.Tests\PairingTest.Tests.csproj", "{E01CEF7B-7B8F-424C-904A-98CB543783FD}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{CBE29D97-3942-407F-AF9D-8DA241BD0DEA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{CBE29D97-3942-407F-AF9D-8DA241BD0DEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{CBE29D97-3942-407F-AF9D-8DA241BD0DEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{CBE29D97-3942-407F-AF9D-8DA241BD0DEA}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E01CEF7B-7B8F-424C-904A-98CB543783FD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E01CEF7B-7B8F-424C-904A-98CB543783FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E01CEF7B-7B8F-424C-904A-98CB543783FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E01CEF7B-7B8F-424C-904A-98CB543783FD}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {FECED454-97E1-4742-B88C-813FCF5B30F4}
-	EndGlobalSection
-EndGlobal
