--- PairingTest/Repositories/Interfaces/IQuestionRepository.cs
@@ -1,9 +1,0 @@
-﻿using PairingTest.Models;
-
-namespace PairingTest.Repositories.Interfaces
-{
-    public interface IQuestionRepository
-    {
-        public IEnumerable<Question> GetQuestions();
-    }
-}
