--- PairingTest/Repositories/QuestionRepository.cs
@@ -1,20 +1,0 @@
-﻿using PairingTest.Models;
-using PairingTest.Repositories.Interfaces;
-
-namespace PairingTest.Repositories
-{
-    public class QuestionRepository : IQuestionRepository
-    {
-        //This method mocks the data which would usually come from a database
-        public IEnumerable<Question> GetQuestions()
-        {
-            return new List<Question>
-            {
-                new Question() { QuestionId = 1, QuestionText = "What is the capital of Cuba?"},
-                new Question() { QuestionId = 2, QuestionText = "What is the capital of France?"},
-                new Question() { QuestionId = 3, QuestionText = "What is the capital of Poland?"},
-                new Question() { QuestionId = 4, QuestionText = "What is the capital of Germany?"}
-            };
-        }
-    }
-}
