--- PairingTest/Views/Questions/Index.cshtml
@@ -1,3 +1,0 @@
-﻿@using PairingTest.Models
-
-<h2>Questions</h2>
