+++ PairingTestApi.sln
@@ -1,0 +1,31 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.2.32616.157
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PairingTestApi", "PairingTestApi\PairingTestApi.csproj", "{CBE29D97-3942-407F-AF9D-8DA241BD0DEA}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PairingTestApi.Tests", "PairingTestApi.Tests\PairingTestApi.Tests.csproj", "{E01CEF7B-7B8F-424C-904A-98CB543783FD}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CBE29D97-3942-407F-AF9D-8DA241BD0DEA}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CBE29D97-3942-407F-AF9D-8DA241BD0DEA}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CBE29D97-3942-407F-AF9D-8DA241BD0DEA}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CBE29D97-3942-407F-AF9D-8DA241BD0DEA}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E01CEF7B-7B8F-424C-904A-98CB543783FD}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E01CEF7B-7B8F-424C-904A-98CB543783FD}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E01CEF7B-7B8F-424C-904A-98CB543783FD}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E01CEF7B-7B8F-424C-904A-98CB543783FD}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {FECED454-97E1-4742-B88C-813FCF5B30F4}
+	EndGlobalSection
+EndGlobal
