+++ PairingTestApi/Controllers/QuestionsController.cs
@@ -1,0 +1,18 @@
+using Microsoft.AspNetCore.Mvc;
+
+namespace PairingTestApi.Controllers
+{
+    [ApiController]
+    [Route("[controller]")]
+    public class QuestionsController : Controller
+    {
+        public QuestionsController()
+        {
+        }
+
+        public IActionResult Index()
+        {
+            return View();
+        }
+    }
+}
