+++ PairingTestApi/Repositories/Interfaces/IQuestionRepository.cs
@@ -1,0 +1,9 @@
+﻿using PairingTestApi.Models;
+
+namespace PairingTestApi.Repositories.Interfaces
+{
+    public interface IQuestionRepository
+    {
+        public IEnumerable<Question> GetQuestions();
+    }
+}
