+++ PairingTestApi/Views/Questions/Index.cshtml
@@ -1,0 +1,3 @@
+﻿@using PairingTestApi.Models
+
+<h2>Questions</h2>
