+++ README.md
@@ -1,12 +1,10 @@
-# .NET Pairing Test
+# Candidate Pairing Test
 
 Take a look over the code in this repo. When you are invited for your interview you will have a 45 minute pair programming session to complete some tasks based on this project.
 
 ### The Code
-
 The application you will be building is an questionnaire and is written using the MVC framework. The solution uses a hardcoded/mocked repository to avoid using a real data source.
 
 It is expected that you will produce:
-
-- Tested maintainable code
-- Clean code and OOP design
+* Tested maintainable code
+* Clean code and OOP design
